[languagedetection] Move privileged detector to separate package
